tests - move findProjectDir in tests/utils
